完成

+++ gd-service/src/main/java/com/project/service/impl/AdminServiceImpl.java
@@ -4,7 +4,11 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.project.constant.RoleConstant;
 import com.project.exception.BusinessExceptionHandler;
 import com.project.mapper.AdminMapper;
+import com.project.mapper.ArticleMapper;
 import com.project.service.AdminService;
+import com.project.service.LikesService;
+import com.project.utils.UserContext;
+import com.project.vo.article.QueryArticleVO;
 import com.project.vo.user.UserVO;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -15,12 +19,17 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
 public class AdminServiceImpl implements AdminService {
     @Resource
     private AdminMapper adminMapper;
+    @Resource
+    private ArticleMapper articleMapper;
+    @Resource
+    private LikesService likesService;
 
     /**
      * 查询普通用户
@@ -73,6 +82,13 @@ public class AdminServiceImpl implements AdminService {
             throw new BusinessExceptionHandler(400, "用户id错误");
         }
 
+        // 获取当前用户的id
+        Long currentUserId = UserContext.getUserId();
+        if (userId.equals(currentUserId)) {
+            log.error("删除单个用户 -----> 不能删除自己");
+            throw new BusinessExceptionHandler(400, "不能删除自己");
+        }
+
         // 执行操作
         return adminMapper.deleteUser(userId);
     }
@@ -86,6 +102,13 @@ public class AdminServiceImpl implements AdminService {
         if (userIdList.isEmpty()) {
             log.error("批量删除用户 -----> 参数错误");
             throw new BusinessExceptionHandler(400, "参数错误");
+        }
+
+        // 获取当前用户的id
+        Long userId = UserContext.getUserId();
+        if (userIdList.contains(userId)) {
+            log.error("批量删除用户 -----> 不能删除自己");
+            throw new BusinessExceptionHandler(400, "不能删除自己");
         }
 
         // 执行操作
@@ -103,8 +126,8 @@ public class AdminServiceImpl implements AdminService {
             throw new BusinessExceptionHandler(400, "参数错误");
         }
 
-        // 查询用户是否已经是管理员
-        String userRole = adminMapper.isAdmin(userId);
+        // 查询用户权限
+        String userRole = adminMapper.getUserRole(userId);
         if (Objects.equals(userRole, RoleConstant.ROLE_ADMIN)) {
             log.error("设置用户为管理员 -----> 当前用户已经是管理员");
             throw new BusinessExceptionHandler(400, "当前用户已经是管理员");
@@ -151,6 +174,60 @@ public class AdminServiceImpl implements AdminService {
         Map<String, Object> map = new HashMap<>();
         map.put("admin", userVOPage.getRecords());
         map.put("total", userVOPage.getTotal());
+        return map;
+    }
+
+    /**
+     * 设置管理员为普通用户
+     */
+    @Override
+    public boolean settingUser(Long userId) {
+        // 校验参数
+        if (userId <= 0) {
+            log.error("设置管理员为普通用户 -----> 参数错误");
+            throw new BusinessExceptionHandler(400, "参数错误");
+        }
+
+        // 查询用户权限
+        String userRole = adminMapper.getUserRole(userId);
+        if (Objects.equals(userRole, RoleConstant.ROLE_USER)) {
+            log.error("设置管理员为普通用户 -----> 当前用户已经是普通用户");
+            throw new BusinessExceptionHandler(400, "当前用户已经是普通用户");
+        }
+
+        // 设置成普通用户
+        return adminMapper.settingUser(userId);
+    }
+
+    /**
+     * 分页查询动态
+     */
+    @Override
+    public Map<String, Object> queryArticle(Integer current, Integer size) {
+        // 查询数据库记录
+        List<QueryArticleVO> queryArticleVOS;
+        Page<QueryArticleVO> queryArticleVOPage;
+        try {
+            Page<QueryArticleVO> page = new Page<>(current, size);
+            queryArticleVOPage = articleMapper.queryArticleByPage(page);
+            queryArticleVOS = queryArticleVOPage.getRecords();
+        } catch (Exception e) {
+            log.error("查询校园动态----->数据库查询失败");
+            throw new RuntimeException(e);
+        }
+
+        // 根据动态id查询点赞数以及自己是否进行点赞
+        if (!queryArticleVOS.isEmpty()) {
+            queryArticleVOS = queryArticleVOS.stream().peek(queryArticleVO -> {
+                // 查询动态点赞数
+                Integer count = likesService.queryLikeCount(queryArticleVO.getId());
+                queryArticleVO.setLikeCount(count);
+            }).collect(Collectors.toList());
+        }
+
+        Map<String, Object> map = new HashMap<>();
+        map.put("article", queryArticleVOS);
+        map.put("total", queryArticleVOPage.getTotal());
         return map;
     }
 }
